Resynced with upx 1.2 branch.
committer: mfx <mfx> 1031577224 +0000

+++ src/s_djgpp2.cpp
@@ -123,7 +123,7 @@ static __inline__
 unsigned short make_cell(screen_t *this, int ch, int attr)
 {
     UNUSED(this);
-    return ((attr & 0xff) << 8) | (ch & 0xff);
+    return (unsigned short) (((attr & 0xff) << 8) | (ch & 0xff));
 }
